requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
openai>=1.55.0
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gunicorn==21.2.0
python-json-logger==2.0.7
redis==5.0.1
celery==5.3.4
sqlalchemy>=2.0.35
# psycopg2-binary==2.9.9  # Uncomment for PostgreSQL support (use SQLite for local dev)
